requireJS, refactoring
