+++ src/filter/filter-type.js
@@ -1,0 +1,11 @@
+'use strict';
+
+/**
+  * @enum
+  * @type {Array.<Object>}
+  */
+module.exports = {
+  POPULAR: 'filter-popular',
+  NEW: 'filter-new',
+  DISCUSSED: 'filter-discussed'
+};
